가챠 시스템 전체 오류 수정 및 서버 정상 동작 복구
Redis 캐시 미적용 문제 해결 및 Warmup 로직 정상화

GachaPool / GachaPoolEntry 컬럼 매핑 오류 수정 (pool_id → PoolId)

EF Core 엔티티 설정 재정비 및 누락된 설정 반영

Proto 응답 매핑 단계에서 발생하던 InvalidCastException 제거

GachaCatalog / PoolDetail API 정상 응답 확인

전체 가챠 기능(배너 조회, 풀 로드, 카탈로그, 뽑기) 문제 없이 작동하도록 안정화 완료

+++ WebServer/Application/Items/Dtos.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 namespace Application.Items
 {
     public sealed record ItemDto(
-        long Id,
+        int Id,
         string Code,
         string Name,
         string Description,
@@ -37,21 +37,21 @@ namespace Application.Items
     );
 
     public sealed record ItemStatDto(
-        long Id,
+        int Id,
         int StatId,
         decimal Value,
         string code
     );
 
     public sealed record ItemEffectDto(
-        long Id,
+        int Id,
         ItemEffectScope Scope,
         JsonDocument Payload,
         short SortOrder
     );
 
     public sealed record ItemPriceDto(
-        long Id,
+        int Id,
         int CurrencyId,
         ItemPriceType PriceType,
         long Price
