+++ WebServer/Domain/Entities/Characters/CharacterPromotionMaterial.cs
@@ -11,12 +11,12 @@ namespace Domain.Entities.Characters
         private CharacterPromotionMaterial() { }
         public int PromotionCharacterId { get; private set; }
         public int PromotionTier { get; private set; }
-        public long ItemId { get; private set; }
+        public int ItemId { get; private set; }
         public int Count { get; private set; }
         public CharacterPromotion Promotion { get; private set; } = null!;
         public Item Item { get; private set; } = null!;
 
-        public static CharacterPromotionMaterial Create(int characterId, int tier, long itemId, int count)
+        public static CharacterPromotionMaterial Create(int characterId, int tier, int itemId, int count)
             => new() { PromotionCharacterId = characterId, PromotionTier = tier, ItemId = itemId, Count = count };
     }
 }
